pyproject.toml
[project]
name = "onnx-inspector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "librosa>=0.11.0",
    "onnx>=1.20.1",
    "onnxruntime>=1.23.2",
    "python-dotenv>=1.2.1",
    "soundfile>=0.13.1",
]
